requirements.txt
Pillow
nose
pyyaml